fix: correctly list php 7.4 as the minimum required version (#334)
Correctly lists php 7.4 as the minimum required version in the
composer.json and README. This is not a breaking change because several
dependencies already refused to install on PHP < 7.4 with the current
version contraints and we have no test coverage for PHP < 7.4. This
commit simply fixes the root-level version contraints to match.